requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi
uvicorn[standard]
python-multipart

# ── Configuration ─────────────────────────────────────────────────────────────
python-dotenv
pydantic

# ── HTTP Client ───────────────────────────────────────────────────────────────
requests

# ── AI / LLM (LangChain + Gemini) ────────────────────────────────────────────
langchain
langchain-core
langchain-google-genai

# ── OCR Model (Donut via HuggingFace Transformers) ───────────────────────────
torch
torchvision
transformers
sentencepiece
protobuf
Pillow

